add sample data and update filterED function
Added a few lines at the end to subset some of the data for testing purposes. Modified the filterED function so that it is now a function of the patient.vars as well. These are defined at the beginning of the script and tell the function what patient variables to bring in.